Add files via upload
weather forecast of 4 days for somebody who wants to go on a trip. The app will tell the client to whether take an umbrella or mask according to the weather forecast. I created a server that was used for calling API from OpenWeatherMap API  using data set of 5 days per 3 hours. I did the front end using Vue js.